Added ability to put your dinos to sleep
Added new "Pet to sleep" multiuse option which maxes the dino's torpor
Added new "Wake Up" multiuse option which reduces dino's torpor to 0 if it was put to sleep using the "Pet to sleep" option.
Added check for damage, if the dino is hit while asleep, it will wake up. It will not be able to be put to sleep or woken up for ten minutes following taking damage.